Move spec generator files, organize whitespace in generator output
